+++ src/interface.ts
@@ -15,8 +15,6 @@ export interface Configuration {
   lightContrast?: string;
   darkWorkbench?: string;
   lightWorkbench?: string;
-  darkSelection?: string;
-  lightSelection?: string;
   darkCursor?: string;
   lightCursor?: string;
   italicKeywords?: boolean;